Remove BaseMessageListHeaderViewModel
