+++ stream-chat-android-ui-common/api/stream-chat-android-ui-common.api
@@ -378,18 +378,6 @@ public final class com/getstream/sdk/chat/view/messages/MessageListItemWrapper {
 	public fun toString ()Ljava/lang/String;
 }
 
-public abstract class com/getstream/sdk/chat/viewmodel/BaseMessageListHeaderViewModel : androidx/lifecycle/ViewModel {
-	public synthetic fun <init> (Ljava/lang/String;Lio/getstream/chat/android/client/ChatClient;Lio/getstream/chat/android/offline/experimental/global/GlobalState;ILkotlin/jvm/internal/DefaultConstructorMarker;)V
-	public final fun getActiveThread ()Landroidx/lifecycle/LiveData;
-	public final fun getChannel ()Landroidx/lifecycle/LiveData;
-	public final fun getChannelState ()Lio/getstream/chat/android/offline/experimental/channel/state/ChannelState;
-	public final fun getMembers ()Landroidx/lifecycle/LiveData;
-	public final fun getOnline ()Landroidx/lifecycle/LiveData;
-	public final fun getTypingUsers ()Landroidx/lifecycle/LiveData;
-	public final fun resetThread ()V
-	public final fun setActiveThread (Lio/getstream/chat/android/client/models/Message;)V
-}
-
 public final class com/getstream/sdk/chat/viewmodel/MessageInputViewModel : androidx/lifecycle/ViewModel {
 	public fun <init> (Ljava/lang/String;)V
 	public fun <init> (Ljava/lang/String;Lio/getstream/chat/android/client/ChatClient;)V
