+++ stream-chat-android-ui-components/api/stream-chat-android-ui-components.api
@@ -2229,9 +2229,16 @@ public final class io/getstream/chat/android/ui/message/list/header/MessageListH
 	public fun toString ()Ljava/lang/String;
 }
 
-public final class io/getstream/chat/android/ui/message/list/header/viewmodel/MessageListHeaderViewModel : com/getstream/sdk/chat/viewmodel/BaseMessageListHeaderViewModel {
-	public fun <init> (Ljava/lang/String;Lio/getstream/chat/android/client/ChatClient;)V
-	public synthetic fun <init> (Ljava/lang/String;Lio/getstream/chat/android/client/ChatClient;ILkotlin/jvm/internal/DefaultConstructorMarker;)V
+public final class io/getstream/chat/android/ui/message/list/header/viewmodel/MessageListHeaderViewModel : androidx/lifecycle/ViewModel {
+	public synthetic fun <init> (Ljava/lang/String;Lio/getstream/chat/android/client/ChatClient;Lio/getstream/chat/android/offline/experimental/global/GlobalState;ILkotlin/jvm/internal/DefaultConstructorMarker;)V
+	public final fun getActiveThread ()Landroidx/lifecycle/LiveData;
+	public final fun getChannel ()Landroidx/lifecycle/LiveData;
+	public final fun getChannelState ()Lio/getstream/chat/android/offline/experimental/channel/state/ChannelState;
+	public final fun getMembers ()Landroidx/lifecycle/LiveData;
+	public final fun getOnline ()Landroidx/lifecycle/LiveData;
+	public final fun getTypingUsers ()Landroidx/lifecycle/LiveData;
+	public final fun resetThread ()V
+	public final fun setActiveThread (Lio/getstream/chat/android/client/models/Message;)V
 }
 
 public final class io/getstream/chat/android/ui/message/list/header/viewmodel/MessageListHeaderViewModelBinding {
